feat: further modularised the code
