+++ greyhoundDashboard/services/navidrome/__init__.py
@@ -1,0 +1,5 @@
+from .service import ServiceNavidrome
+
+__all__ = [
+    "ServiceNavidrome",
+]
