+++ greyhoundDashboard/services/portainer/__init__.py
@@ -1,0 +1,5 @@
+from .service import ServicePortainer
+
+__all__ = [
+    "ServicePortainer",
+]
