+++ greyhoundDashboard/services/vikunja/__init__.py
@@ -1,0 +1,5 @@
+from .service import ServiceVikunja
+
+__all__ = [
+    "ServiceVikunja",
+]
